Added some kernel math funcs, removed some FindStmts
Added two kernel math functions that are required for one application.
Plans to add more will come. Also removed some FindStmts and replaced
them by rewriting the expression. This provides a slightly better
mechanism. In the future, try to query the expression type to find
the required one and rewrite that expression into a string.